OPIK-2946 PR comments
