+++ apps/opik-backend/src/main/java/com/comet/opik/api/CreateDatasetItemsFromTracesRequest.java
@@ -1,6 +1,6 @@
 package com.comet.opik.api;
 
-import com.comet.opik.domain.TraceEnrichmentService;
+import com.comet.opik.domain.TraceEnrichmentOptions;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -15,5 +15,5 @@ import java.util.UUID;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public record CreateDatasetItemsFromTracesRequest(
         @NotEmpty(message = "trace_ids must not be empty") @Schema(description = "Set of trace IDs to add to the dataset", requiredMode = Schema.RequiredMode.REQUIRED) Set<UUID> traceIds,
-        @NotNull(message = "enrichment_options must not be null") @Schema(description = "Options for enriching trace data", requiredMode = Schema.RequiredMode.REQUIRED) TraceEnrichmentService.TraceEnrichmentOptions enrichmentOptions) {
+        @NotNull(message = "enrichment_options must not be null") @Schema(description = "Options for enriching trace data", requiredMode = Schema.RequiredMode.REQUIRED) TraceEnrichmentOptions enrichmentOptions) {
 }
