+++ apps/opik-backend/src/main/java/com/comet/opik/domain/SpanEnrichmentOptions.java
@@ -1,0 +1,15 @@
+package com.comet.opik.domain;
+
+import lombok.Builder;
+
+/**
+ * Options for span enrichment specifying which metadata to include.
+ */
+@Builder(toBuilder = true)
+public record SpanEnrichmentOptions(
+        boolean includeTags,
+        boolean includeFeedbackScores,
+        boolean includeComments,
+        boolean includeUsage,
+        boolean includeMetadata) {
+}
