+++ apps/opik-backend/src/main/java/com/comet/opik/domain/SpanEnrichmentService.java
@@ -1,24 +1,17 @@
 package com.comet.opik.domain;
 
-import com.comet.opik.api.Span;
-import com.comet.opik.utils.EnrichmentUtils;
-import com.comet.opik.utils.JsonUtils;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.opentelemetry.instrumentation.annotations.WithSpan;
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
-import lombok.Builder;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
 
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 /**
  * Service responsible for enriching span data with additional metadata
@@ -49,80 +42,9 @@ public class SpanEnrichmentService {
 
         log.info("Enriching '{}' spans with options '{}'", spanIds.size(), options);
 
-        // Fetch all spans
+        // Fetch all spans and enrich them reactively
         return spanService.getByIds(spanIds)
-                .collectList()
-                .flatMap(spans -> {
-                    Map<UUID, Map<String, JsonNode>> enrichedSpans = spans.stream()
-                            .collect(Collectors.toMap(
-                                    Span::id,
-                                    span -> enrichSpanData(span, options)));
-
-                    return Mono.just(enrichedSpans);
-                });
-    }
-
-    private Map<String, JsonNode> enrichSpanData(
-            @NonNull Span span,
-            @NonNull SpanEnrichmentOptions options) {
-
-        ObjectNode enrichedData = JsonUtils.getMapper().createObjectNode();
-
-        // Always include input and output
-        enrichedData.set("input", JsonUtils.getMapper().valueToTree(span.input()));
-        Optional.ofNullable(span.output())
-                .ifPresent(output -> enrichedData.set("expected_output", JsonUtils.getMapper().valueToTree(output)));
-
-        // Include tags if requested
-        if (options.includeTags()) {
-            Optional.ofNullable(span.tags())
-                    .filter(tags -> !tags.isEmpty())
-                    .ifPresent(tags -> enrichedData.set("tags", JsonUtils.getMapper().valueToTree(tags)));
-        }
-
-        // Include feedback scores if requested
-        if (options.includeFeedbackScores()) {
-            Optional.ofNullable(span.feedbackScores())
-                    .filter(scores -> !scores.isEmpty())
-                    .ifPresent(scores -> enrichedData.set("feedback_scores",
-                            EnrichmentUtils.buildFeedbackScoresNode(scores)));
-        }
-
-        // Include comments if requested
-        if (options.includeComments()) {
-            Optional.ofNullable(span.comments())
-                    .filter(comments -> !comments.isEmpty())
-                    .ifPresent(comments -> enrichedData.set("comments",
-                            EnrichmentUtils.buildCommentsNode(comments)));
-        }
-
-        // Include usage if requested
-        if (options.includeUsage()) {
-            Optional.ofNullable(span.usage())
-                    .filter(usage -> !usage.isEmpty())
-                    .ifPresent(usage -> enrichedData.set("usage", JsonUtils.getMapper().valueToTree(usage)));
-        }
-
-        // Include metadata if requested
-        if (options.includeMetadata()) {
-            Optional.ofNullable(span.metadata())
-                    .ifPresent(metadata -> enrichedData.set("metadata", JsonUtils.getMapper().valueToTree(metadata)));
-        }
-
-        // Convert ObjectNode to Map<String, JsonNode>
-        return enrichedData.properties().stream()
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-    }
-
-    /**
-     * Options for span enrichment specifying which metadata to include.
-     */
-    @Builder(toBuilder = true)
-    public record SpanEnrichmentOptions(
-            boolean includeTags,
-            boolean includeFeedbackScores,
-            boolean includeComments,
-            boolean includeUsage,
-            boolean includeMetadata) {
+                .map(span -> Map.entry(span.id(), SpanEnrichmentMapper.enrichSpanData(span, options)))
+                .collectMap(Map.Entry::getKey, Map.Entry::getValue);
     }
 }
